requirements.txt
langgraph==0.2.50
langchain-google-genai==2.0.7
langchain-core>=0.3.15,<0.4.0
google-generativeai==0.8.3
streamlit==1.40.0
python-dotenv==1.0.0
pandas==2.2.0
requests==2.31.0
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.118.0
